refactor: move component to ui as it is used in ui and in the home feature
